refactor : Request, Response
- Refactoring on Request, Response
