+++ waldreg/application-layer/schedule-controller/src/main/java/org/waldreg/controller/schedule/ScheduleController.java
@@ -1,0 +1,4 @@
+package org.waldreg.controller.schedule;
+
+public class ScheduleController{
+}
